requirements.txt
opencv-python==4.8.1.78
openai==1.3.7
google-generativeai==0.3.2
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.24.3